requirements.txt
streamlit
pymupdf 
python-docx 
pytesseract 
pillow
opneai
# pyheif
pillow-heif
dotenv
fpdf
